LayoutTests/fast/css-grid-layout/breadth-size-resolution-grid.html

<!DOCTYPE html>
<html>
<script>
if (window.testRunner)
    testRunner.overridePreference("WebKitCSSGridLayoutEnabled", 1);
</script>
<link href="resources/grid.css" rel="stylesheet">
<style>
.grid {
    -webkit-grid-template-columns: 100px 10vw;
    -webkit-grid-template-rows: 10vh 100px;
}

.firstRowFirstColumn {
    width: 100%;
    height: 100%;
}

.firstRowSecondColumn {
    width: 100%;
    height: 100%;
}

.secondRowFirstColumn {
    width: 100%;
    height: 100%;
}

.secondRowSecondColumn {
    width: 100%;
    height: 100%;
}
</style>
<script src="../../resources/check-layout.js"></script>
<body onload="checkLayout('.grid')">

<p>Test that specifying the track breadth sizes works properly with all the allowed length types and in different writing modes.</p>

<div class="grid">
    <div class="firstRowFirstColumn" data-expected-width="100" data-expected-height="60"></div>
    <div class="firstRowSecondColumn" data-expected-width="80" data-expected-height="60"></div>
    <div class="secondRowFirstColumn" data-expected-width="100" data-expected-height="100"></div>
    <div class="secondRowSecondColumn" data-expected-width="80" data-expected-height="100"></div>
</div>

<div class="grid verticalRL">
    <div class="firstRowFirstColumn" data-expected-width="60" data-expected-height="100"></div>
    <div class="firstRowSecondColumn" data-expected-width="60" data-expected-height="80"></div>
    <div class="secondRowFirstColumn" data-expected-width="100" data-expected-height="100"></div>
    <div class="secondRowSecondColumn" data-expected-width="100" data-expected-height="80"></div>
</div>

<div class="grid">
     <div class="firstRowFirstColumn" class="verticalRL" data-expected-width="100" data-expected-height="60"></div>
     <div class="firstRowSecondColumn" data-expected-width="80" data-expected-height="60"></div>
     <div class="secondRowFirstColumn" class="verticalRL" data-expected-width="100" data-expected-height="100"></div>
     <div class="secondRowSecondColumn" data-expected-width="80" data-expected-height="100"></div>
</div>

</body>
</html>
